refactor env modules and adjust camera
